+++ src/db.rs
@@ -11,9 +11,3 @@ pub fn fetchone<T: FromSql>(
         conn.query_row(sql, [], |row| row.get(0)).optional()
     }
 }
-/// open media db
-pub fn open_media_db<P: AsRef<Path>>(path: P) -> Result<Connection> {
-    let db = Connection::open(path)?;
-    db.execute_batch(include_str!("schema.sql"))?;
-    Ok(db)
-}
